PlayerListener separation of concerns refactoring.
